requirements.txt
alembic==1.0.0
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
clickclick==1.2.2
connexion==2.0.2
dominate==2.3.4
envparse==0.2.0
Flask==1.0.2
Flask-Cors==3.0.4
Flask-DotEnv==0.1.1
fleep==1.0.1
gevent==1.3.6
greenlet==0.4.15
gunicorn==19.9.0
idna==2.6
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.2.0
openapi-spec-validator==0.2.4
passlib==1.7.1
pluggy==0.8.0
prison==0.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.3
pytest==3.8.1
pytest-ordering==0.6
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
PyYAML==3.13
requests==2.20.1
six==1.11.0
spec-synthase==0.1.9
SQLAlchemy==1.2.7
swagger-parser==1.0.1
swagger-spec-validator==2.4.1
swagger-tester==0.2.12
swaggertools==0.0.4
urllib3==1.22
uWSGI==2.0.17.1
Werkzeug==0.14.1
